The function plots the V-fold cross-validated risk estimates for the super learner, the discrete super learner and each algorithm in the library. By default the estimates will be sorted and include an asymptotic 95% confidence interval.

# S3 method for CV.SuperLearner
plot(x, package = "ggplot2", constant = qnorm(0.975), sort = TRUE, ...)

Arguments

x

The output from CV.SuperLearner.

package

Either "ggplot2" or "lattice". The package selected must be available.

constant

A numeric value. The confidence interval is defined as p +/- constant * se, where p is the point estimate and se is the standard error. The default is the quantile of the standard normal corresponding to a 95% CI.

sort

Logical. Should the rows in the plot be sorted from the smallest to the largest point estimate. If FALSE, then the order is super learner, discrete super learner, then the estimators in SL.library.

…

Additional arguments for summary.CV.SuperLearner

Details

see summary.CV.SuperLearner for details on how the estimates are computed

Value

Returns the plot (either a ggplot2 object (class ggplot) or a lattice object (class trellis))

See also

summary.CV.SuperLearner and CV.SuperLearner